Splitting js files into smaller partitions, correcting json-parsing issues in returns from controllers
