+++ src/main/webapp/resources/js/portal.js
@@ -17,7 +17,6 @@ angular.module("scrumApp", [])
 })
 
 .directive("scrumBody", function() {
-	
 	return {
 		restrict: 'E',
 		scope: { fragment: "=" },
@@ -28,80 +27,6 @@ angular.module("scrumApp", [])
 	        }
 	      });
 	    },
-		 
 	    template: '<ng-include src="dynamicTemplateUrl"></ng-include>'
 	};
-})
-
-.controller("loginCtrl", function($scope, dataService) {
-	$scope.error = false;
-	$scope.errorMessage = "";
-	$scope.submitted = false;
-	$scope.bu = {
-		username: "",
-		password: ""
-	};
-	
-	$scope.login = function(bu) {
-		if (!$scope.submitted) {
-			console.log("Preparing to login");
-			$scope.submitted = true;
-			dataService.login(bu, $scope);
-		}
-	}
-})
-
-.service("dataService", function($http, $window, $rootScope) {
-	this.login = function(bu, $scope) {
-		var promise = $http.post("login", bu).then(
-				function success(response) {
-					$rootScope.$emit("toggleLogin", {});
-					$rootScope.$emit("updateFragment", "dropdown");
-				},
-				function error(response) {
-					if (response.status == 409) {
-						$scope.errorMessage = response.data[0];
-					} else {
-						$scope.errorMessage = "Error adding user.";
-					}
-					$scope.error = true;
-					$scope.submitted = false;
-				}
-		)
-	}
-	
-	this.isLoggedIn = function() {
-		var promise = $http.get("ajaxIsLoggedIn").then(
-				function success(response) {
-					return true;
-				},
-				function error(response) {
-					return false;
-				}
-		)
-	}
-	
-	this.logout = function() {
-		var promise = $http.get("logout").then(
-				function success(response) {
-					$rootScope.$emit("toggleLogin", {});
-					$rootScope.$emit("updateFragment", "login");
-				},
-				function error(response) {
-					console.log(response.data[0]);
-				}
-		)
-	}
-	
-	this.getBoards = function() {
-		var promise = $http.get("dropdown").then(
-				function success(response) {
-					$rootScope.$emit("toggleLogin", {});
-					$rootScope.$emit("updateFragment", "login");
-				},
-				function error(response) {
-					console.log(response.data[0]);
-				}
-		)
-	}
 })
