+++ src/main/webapp/resources/portal.html
@@ -42,6 +42,8 @@
 	</div>
 
 	<script type="text/javascript" src="resources/js/portal.js"></script>
+	<script type="text/javascript" src="resources/js/dataService.js"></script>
 	<script type="text/javascript" src="resources/js/navbarCtrl.js"></script>
+	<script type="text/javascript" src="resources/js/loginCtrl.js"></script>
 </body>
 </html>
